# api.swagger.yaml
swagger: "2.0"
info:
  title: Blog Server.
  description: Entry point for the application.
  version: 1.0.0
  contact:
    name: HardDie
    url: https://github.com/HardDie
    email: oleg1995sysoev@yandex.ru
...
basePath: /
schemes:
  - https
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /api/v1/posts:
    get:
      summary: Get a list of posts for the current user
      operationId: Post_List
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/gatewayListResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: limit
          in: query
          required: false
          type: integer
          format: int32
        - name: page
          in: query
          required: false
          type: integer
          format: int32
        - name: query
          in: query
          required: false
          type: string
      tags:
        - Post
    post:
      summary: Post creation form
      operationId: Post_Create
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/gatewayCreateResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/gatewayCreateRequest'
      tags:
        - Post
  /api/v1/posts/feed:
    get:
      summary: Get feed
      operationId: Post_Feed
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/gatewayFeedResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: limit
          in: query
          required: false
          type: integer
          format: int32
        - name: page
          in: query
          required: false
          type: integer
          format: int32
        - name: query
          in: query
          required: false
          type: string
      tags:
        - Post
  /api/v1/posts/{id}:
    get:
      summary: Get public post
      operationId: Post_PublicGet
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/gatewayPublicGetResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: id
          in: path
          required: true
          type: string
          format: int64
      tags:
        - Post
    put:
      summary: Edit post form
      operationId: Post_Edit
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/gatewayEditResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: id
          in: path
          required: true
          type: string
          format: int64
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/PostEditBody'
      tags:
        - Post
...
definitions:
  PostEditBody:
    type: object
    properties:
      title:
        type: string
      short:
        type: string
      body:
        type: string
      tags:
        type: array
        items:
          type: string
      isPublished:
        type: boolean
  gatewayCreateRequest:
    type: object
    properties:
      title:
        type: string
      short:
        type: string
      body:
        type: string
      tags:
        type: array
        items:
          type: string
      isPublished:
        type: boolean
  gatewayCreateResponse:
    type: object
    properties:
      data:
        $ref: '#/definitions/gatewayPostObject'
  gatewayEditResponse:
    type: object
    properties:
      data:
        $ref: '#/definitions/gatewayPostObject'
  gatewayFeedResponse:
    type: object
    properties:
      data:
        type: array
        items:
          type: object
          $ref: '#/definitions/gatewayPostObject'
...
  gatewayListResponse:
    type: object
    properties:
      data:
        type: array
        items:
          type: object
          $ref: '#/definitions/gatewayPostObject'
...
  gatewayPostObject:
    type: object
    properties:
      id:
        type: string
        format: int64
      userId:
        type: string
        format: int64
      user:
        $ref: '#/definitions/gatewayPublicUserObject'
      title:
        type: string
      short:
        type: string
      body:
        type: string
      tags:
        type: array
        items:
          type: string
      isPublished:
        type: boolean
      createdAt:
        type: string
        format: date-time
...
  gatewayPublicGetResponse:
    type: object
    properties:
      data:
        $ref: '#/definitions/gatewayPostObject'
  gatewayPublicUserObject:
    type: object
    properties:
      id:
        type: string
        format: int64
      displayedName:
        type: string
      invitedByUserId:
        type: string
        format: int64
      createdAt:
        type: string
        format: date-time
...
  protobufAny:
    type: object
    properties:
      '@type':
        type: string
    additionalProperties: {}
  rpcStatus:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      details:
        type: array
        items:
          type: object
          $ref: '#/definitions/protobufAny'
securityDefinitions:
  bearer:
    type: apiKey
    description: 'Authentication token, prefixed by Bearer: Bearer <token>'
    name: Authorization
    in: header
security:
  - bearer: []